+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log for cohesity/community-automation-samples
 
+## 2024-05-01
+
+* [oracleLogDeletionDaysReport.py](https://github.com/cohesity/community-automation-samples/tree/main/oracle/python/oracleLogDeletionDaysReport) [`Update`] updated to gracefully handle edge cases
+
 ## 2024-04-30
 
 * [restoreOracle-v2.ps1](https://github.com/cohesity/community-automation-samples/tree/main/oracle/powershell/restoreOracle-v2) [`Fix`] fixed recovery error